[101] moved categories and course selection to the CommonInputEventForm.js
